+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGLES/types.cpp
@@ -1,0 +1,9 @@
+#include "Graphics_Systems/OpenGL-Common/version.h"
+
+namespace enigma {
+
+int getGLTypeSize(GLuint type) {
+  return OpenGL::getGLTypeSize(type);
+}
+
+} //namespace enigma
