+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGLES3/std.cpp
@@ -6,4 +6,8 @@ const gl_profile_type graphics_opengl_profile = gl_profile_es;
 const unsigned int gl_major = 3, gl_minor = 2; 
 const std::string shader_header = "#version 320 es\nprecision mediump float;\n";
 
+GLenum shadertypes[5] = {
+  GL_VERTEX_SHADER, GL_FRAGMENT_SHADER, GL_TESS_CONTROL_SHADER, GL_TESS_EVALUATION_SHADER, GL_GEOMETRY_SHADER
+};
+
 } // namespace enigma
