+++ ENIGMAsystem/SHELL/Platforms/SDL/Makefile
@@ -5,6 +5,10 @@ endif
 SOURCES += $(wildcard Platforms/SDL/*.cpp) Platforms/General/POSIX/Utility.cpp
 override CXXFLAGS += $(shell $(PKG-CONFIG) --cflags sdl2) -DENIGMA_PLATFORM_SDL
 
+ifneq ($(TARGET-PLATFORM), Android)
+	SOURCES += Platforms/General/POSIX/shell.cpp
+endif
+
 ifeq ($(TARGET-PLATFORM), Windows)
 	SOURCES += $(wildcard Bridges/SDL-Win32/*.cpp) Platforms/Win32/WINDOWSfilemanip.cpp
 override LDLIBS += $(shell $(PKG-CONFIG) --libs --static sdl2)
